add client settings, add supporing i18n for telegram bot, add print statistic mode and more

+++ src/main/java/ru/akvine/fitstats/controllers/rest/dto/diet/DietDisplayInfoDto.java
@@ -6,27 +6,27 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class DietDisplayInfoDto {
-    private int maxProteins;
+    private double maxProteins;
 
-    private int maxFats;
+    private double maxFats;
 
-    private int maxCarbohydrates;
+    private double maxCarbohydrates;
 
-    private int maxCalories;
+    private double maxCalories;
 
-    private int currentProteins;
+    private double currentProteins;
 
-    private int currentFats;
+    private double currentFats;
 
-    private int currentCarbohydrates;
+    private double currentCarbohydrates;
 
-    private int currentCalories;
+    private double currentCalories;
 
-    private int remainingProteins;
+    private double remainingProteins;
 
-    private int remainingFats;
+    private double remainingFats;
 
-    private int remainingCarbohydrates;
+    private double remainingCarbohydrates;
 
-    private int remainingCalories;
+    private double remainingCalories;
 }
